+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # uwsgiconf changelog
 
+### Unreleased
+* ** Django contrib. Use qualname for task function name.
+
 ### v2.3.0 [2025-12-06]
 * ++ Django contrib. Task details pages now show additional information.
 * ** Django contrib. Fixed task forced task run missing tasks message (fixes #16).