+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # uwsgiconf changelog
 
-### Unreleased
+### v2.3.1 [2026-04-25]
 * ** Django contrib. Use qualname for task function name.
 * ** Runtime. Fixed zeros handling for 'register_cron' params (closes #18).